Simple Logger implemented

+++ src/Types.h
@@ -6,7 +6,7 @@
 typedef bool (*ErrorCallbackFunctionType)(const char *errorMessage);
 
 typedef uint16_t led_index_t;
-typedef uint8_t index_id_t;
+typedef uint16_t index_id_t;
 
 class LedStripMode;
 class Animation;
